% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autism}
\alias{autism}
\title{Adult Autism Spectrum Disorder (ASD) screening}
\format{
A data frame with 20 variables and 704 observations.
 \describe{
     \item{row}{Adults.}
     \item{col}{15 binary (and 5 additional) features.}
     \item{groups}{Normal and ASD-diagnosed adults.} 
 }
}
\usage{
autism
}
\description{
This dataset has 15 binary (and 5 additional) features on 515 normal and 189 ASD-diagnosed adults.
}
\keyword{datasets}
