% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex}
\alias{gtex}
\title{RNA sequences of human tissues}
\format{
A data frame with 20242 variables and 2254 observations.
 \describe{
     \item{row}{RNA sequences.}
     \item{col}{Gene expression levels, in FPKM (Fragments per Kilobase of transcripts per Million).}
     \item{groups}{human organs: esophagus, colon, thyroid, lung, breast, stomach, liver and prostate.} 
 }
}
\usage{
gtex
}
\description{
This dataset consists of gene expression levels of RNA sequences from eight human organs, and it has 20242 discrete features, however some of them have so many discrete values, and the probabilities for these discrete values are almost equal and very small. So, these features have approximately continuous CDFs and can be considered to be continuous. 
}
\keyword{datasets}
