% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scripts}
\alias{scripts}
\title{Indic scripts}
\format{
A data frame with 116 variables and 747 observations.
 \describe{
     \item{row}{Handwritten scripts.}
     \item{col}{Features.}
     \item{groups}{Indic languages: Bangla, Gurmukhi, Gujarati, Kannada, Malayalam and Urdu.} 
 }
}
\usage{
scripts
}
\description{
This dataset is on 116 different features from handwritten scripts of 6 Indic languages, where 5 languages from 4 regions, namely Bangla (from the east),   Gurmukhi (north), Gujarati (west), Kannada and Malayalam (languages from the neighboring southern states of Karnataka and Kerala) and a sixth language (Urdu, with a distinct Persian script).
}
\keyword{datasets}
