% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senators}
\alias{senators}
\title{Voting records of US senators}
\format{
A data frame with 542 variables and 100 observations.
 \describe{
     \item{row}{Senators' vote.}
     \item{col}{Bills.}
     \item{groups}{Republican and Democratic.} 
 }
}
\usage{
senators
}
\description{
This dataset contains the 108th US Congress had 55 Republican and 45 Democratic (including 1 independent in the Democratic caucus) senators' vote on 542 bills.
}
\keyword{datasets}
