% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spect}
\alias{spect}
\title{SPECT heart dataset}
\format{
A data frame with 542 variables and 100 observations.
 \describe{
     \item{row}{Patients.}
     \item{col}{Binary attributes.}
     \item{groups}{Normal and abnormal.} 
 }
}
\usage{
spect
}
\description{
This dataset has 22 binary attributes that summarize cardiac Single Proton Emission Computed Tomography (SPECT) images of 55 normal and 212 abnormal patients. Each image was summarized by means of 44 continuous features that was further processed to obtain 22 binary features.
}
\keyword{datasets}
